sandbox/gamemaster/requirements.txt
flask
flask-socketio
eventlet
docker-compose
paramiko